server/requirements.txt
fastapi
uvicorn
sqlalchemy
httpx
passlib
python-dotenv
python-jose
python-socketio
fastapi-socketio
bcrypt
pydantic
mysql-connector-python
faker
faker[providers]
sio
python-socketio